Simplify, improve performance.
Factored the directory walking code out of main so that it can return a
Result, which in turn lets us use `try!` instead of manually
unwrapping/aborting.

Removed some unnecessary conversions and patterns.

Switched the map update to the "accumulator map" pattern discussed in
the docs for std::collections.

Performance is up by 34% for my test dataset, making it marginally
faster than the Ruby version.